# Conflicts:
#   frontend/src/views/main/components/profile-dialog.vue

+++ backend/src/main/java/com/ssafy/api/controller/UserController.java
@@ -3,32 +3,30 @@ package com.ssafy.api.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.ssafy.api.request.UserLoginPostReq;
+import com.ssafy.api.request.UserFindPostReq;
 import com.ssafy.api.request.UserModifyPutReq;
 import com.ssafy.api.request.UserRegisterPostReq;
-import com.ssafy.api.response.UserLoginPostRes;
 import com.ssafy.api.response.UserRes;
 import com.ssafy.api.service.UserService;
 import com.ssafy.common.auth.SsafyUserDetails;
 import com.ssafy.common.model.response.BaseResponseBody;
-import com.ssafy.common.util.JwtTokenUtil;
 import com.ssafy.db.entity.User;
-import com.ssafy.db.repository.UserRepositorySupport;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
+import lombok.RequiredArgsConstructor;
 import springfox.documentation.annotations.ApiIgnore;
 
 /**
@@ -36,11 +34,16 @@ import springfox.documentation.annotations.ApiIgnore;
  */
 @Api(value = "유저 API", tags = {"User"})
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/api/v1/users")
 public class UserController {
 	// 정태현 210719 테스트
 	@Autowired
 	UserService userService;
+	
+//	private final JavaMailSender javaMailSender;
+	
+//	MailBodyUtil mailBodyUtil = new MailBodyUtil();
 	
 	@PostMapping("/create")
 	@ApiOperation(value = "회원 가입", notes = "<strong>아이디와 패스워드</strong>를 통해 회원가입 한다.") 
@@ -54,7 +57,7 @@ public class UserController {
 			@RequestBody @ApiParam(value="회원가입 정보", required = true) UserRegisterPostReq registerInfo) {
 		
 		
-		System.out.println("들어왔다아아앙");
+		System.out.println("회원가입");
 		//임의로 리턴된 User 인스턴스. 현재 코드는 회원 가입 성공 여부만 판단하기 때문에 굳이 Insert 된 유저 정보를 응답하지 않음.
 		User user = userService.createUser(registerInfo);
 		if(user == null)
@@ -65,6 +68,35 @@ public class UserController {
 		
 		
 		return ResponseEntity.status(200).body(BaseResponseBody.of(200, "Success"));
+	}
+	@PostMapping("/findid")
+	@ApiOperation(value = "아이디 찾기", notes = "회원의 아이디를 찾는다") 
+    @ApiResponses({
+        @ApiResponse(code = 200, message = "성공"),
+        @ApiResponse(code = 401, message = "인증 실패"),
+        @ApiResponse(code = 404, message = "사용자 없음"),
+        @ApiResponse(code = 500, message = "서버 오류")
+    })
+	public ResponseEntity<String> findUser(@RequestBody @ApiParam(value="아이디 찾기", required = true) UserFindPostReq findUser) {
+		/**
+		 * 요청 헤더 액세스 토큰이 포함된 경우에만 실행되는 인증 처리이후, 리턴되는 인증 정보 객체(authentication) 통해서 요청한 유저 식별.
+		 * 액세스 토큰이 없이 요청하는 경우, 403 에러({"error": "Forbidden", "message": "Access Denied"}) 발생.
+		 */
+		
+		
+		System.out.println("아이디 찾기");
+		User user = userService.findUserId(findUser.getName(), findUser.getEmail());
+
+		String userId = null;
+		
+		if(user!=null) {
+			System.out.println("아이디 찾았다");
+			userId = user.getUserId();
+			return ResponseEntity.status(200).body(userId);
+		}
+
+		System.out.println("아이디 못찾았다");
+		return ResponseEntity.status(400).body(null);
 	}
 	
 	@GetMapping("/me")
@@ -81,15 +113,17 @@ public class UserController {
 		 * 액세스 토큰이 없이 요청하는 경우, 403 에러({"error": "Forbidden", "message": "Access Denied"}) 발생.
 		 */
 		System.out.println("회원정보 조회");
+		System.out.println(authentication);
 		SsafyUserDetails userDetails = (SsafyUserDetails)authentication.getDetails();
 		String userId = userDetails.getUsername();
 		User user = userService.getUserByUserId(userId);
-		System.out.println(user.getPassword());
+		System.out.println(user.toString());
+//		System.out.println(user.getPassword());
 		
 		return ResponseEntity.status(200).body(UserRes.of(user));
 	}
 
-	@DeleteMapping("/delete")
+	@DeleteMapping("/{userId}")
 	@ApiOperation(value = "회원 탈퇴", notes = "로그인한 회원 본인의 정보를 삭제한다.") 
 	@ApiResponses({
 		@ApiResponse(code = 200, message = "성공"),
@@ -97,21 +131,24 @@ public class UserController {
 		@ApiResponse(code = 404, message = "사용자 없음"),
 		@ApiResponse(code = 500, message = "서버 오류")
 	})
-	public ResponseEntity<UserRes> deleteUserInfo(@ApiIgnore Authentication authentication) {
+	public ResponseEntity<? extends BaseResponseBody> deleteUserInfo(@ApiIgnore Authentication authentication, @PathVariable("userId") String userId) {
 		/**
 		 * 요청 헤더 액세스 토큰이 포함된 경우에만 실행되는 인증 처리이후, 리턴되는 인증 정보 객체(authentication) 통해서 요청한 유저 식별.
 		 * 액세스 토큰이 없이 요청하는 경우, 403 에러({"error": "Forbidden", "message": "Access Denied"}) 발생.
 		 */
+		System.out.println("회원 탈퇴");
 		SsafyUserDetails userDetails = (SsafyUserDetails)authentication.getDetails();
-		String userId = userDetails.getUsername();
-		User user = userService.getUserByUserId(userId);
-		Long Id = user.getId();
-		userService.deleteById(Id);
-		
-		return ResponseEntity.status(200).body(UserRes.of(user));
+		String getUserId = userDetails.getUsername();
+		if(getUserId.equals(userId)) {
+			User user = userService.getUserByUserId(userId);
+			Long Id = user.getId();
+			userService.deleteById(Id);
+			return ResponseEntity.status(200).body(BaseResponseBody.of(200, "Success"));
+		}
+		return ResponseEntity.status(500).body(BaseResponseBody.of(500, "fail"));
 	}
 	
-	@PutMapping("/modify")
+	@PatchMapping("/{userId}")
 	@ApiOperation(value = "회원 정보 수정", notes = "로그인한 회원 본인의 정보를 수정한다.") 
 	@ApiResponses({
 		@ApiResponse(code = 200, message = "성공"),
@@ -119,22 +156,31 @@ public class UserController {
 		@ApiResponse(code = 404, message = "사용자 없음"),
 		@ApiResponse(code = 500, message = "서버 오류")
 	})
-	public ResponseEntity<UserRes> modifyUserInfo(@ApiIgnore Authentication authentication,
+	public ResponseEntity<? extends BaseResponseBody> modifyUserInfo(@ApiIgnore Authentication authentication,@PathVariable("userId") String userId,
 			@RequestBody @ApiParam(value="회원 정보 수정", required = true) UserModifyPutReq modifyInfo) {
 		/**
 		 * 요청 헤더 액세스 토큰이 포함된 경우에만 실행되는 인증 처리이후, 리턴되는 인증 정보 객체(authentication) 통해서 요청한 유저 식별.
 		 * 액세스 토큰이 없이 요청하는 경우, 403 에러({"error": "Forbidden", "message": "Access Denied"}) 발생.
 		 */
+		
+		System.out.println("회원정보수정");
+		
 		SsafyUserDetails userDetails = (SsafyUserDetails)authentication.getDetails();
-		String userId = userDetails.getUsername();
-		User user = userService.getUserByUserId(userId);
-		Long Id = user.getId();
+		String getUserId = userDetails.getUsername();
 		
-		User modify_user = userService.modifyUser(modifyInfo, userId, Id);
+		System.out.println(getUserId);
+		System.out.println(userId);
 		
+		if(getUserId.equals(userId)) {
+			User getUser = userService.getUserByUserId(userId);
+			Long id = getUser.getId();
+			User modifyUser = userService.modifyUser(modifyInfo, userId, id);
+			
+			return ResponseEntity.status(200).body(BaseResponseBody.of(200, "Success"));
+		}
 		
+		return ResponseEntity.status(500).body(BaseResponseBody.of(500, "수정 불가"));
 		
-		return ResponseEntity.status(200).body(UserRes.of(modify_user));
 	}
